backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
akshare
pillow
pytesseract
psycopg2-binary
